Add test case for https://bugs.swift.org/browse/SR-8643 which is now fixed
