+++ test/SILGen/keypaths_multi_file.swift
@@ -1,5 +1,11 @@
-// RUN: %target-swift-emit-silgen -module-name keypaths -primary-file %s %S/Inputs/keypaths_multi_file_b.swift | %FileCheck %s
-// RUN: %target-swift-emit-silgen -module-name keypaths %s -primary-file %S/Inputs/keypaths_multi_file_b.swift | %FileCheck --check-prefix=DEFINITION %s
+// RUN: %empty-directory(%t)
+
+// RUN: %target-build-swift -module-name keypaths_resilient -emit-module -Xfrontend -enable-key-path-resilience %S/Inputs/keypaths_multi_file_c.swift -emit-module-path %t/keypaths_resilient.swiftmodule
+
+// RUN: %target-swift-emit-silgen -module-name keypaths -primary-file %s %S/Inputs/keypaths_multi_file_b.swift -I %t | %FileCheck %s
+// RUN: %target-swift-emit-silgen -module-name keypaths %s -primary-file %S/Inputs/keypaths_multi_file_b.swift -I %t | %FileCheck --check-prefix=DEFINITION %s
+
+import keypaths_resilient
 
 func foo(x: Int) -> KeyPath<A, Int> {
   switch x {
@@ -18,3 +24,22 @@ func foo(x: Int) -> KeyPath<A, Int> {
 // A.subscript setter
 // CHECK-LABEL: sil hidden_external @$S8keypaths1AVyS2icis
 // DEFINITION-LABEL: sil hidden @$S8keypaths1AVyS2icis
+
+func bar<T>(_: T) {
+  _ = \C<T>.b
+  _ = \C<T>[0]
+
+  _ = \D<T>.b
+  _ = \D<T>[0]
+
+  _ = \P.b
+
+  // FIXME: crashes
+  // _ = \P[0]
+}
+
+// https://bugs.swift.org/browse/SR-8643
+class MM<T: P> : PP {}
+func foo<T3: BB, T4: MM<T3>>(t: T3, u: T4) {
+  let _ = \CC<T4>.x
+}
